# gateway/openapi.yaml
openapi: 3.0.0
info:
  title: GetawayAgentes Gateway API
  version: 1.0.0
  description: API central para orquestar y gestionar agentes IA
  contact:
    name: Padilla585
servers:
  - url: https://getaway-gateway.alejandra-app.workers.dev
    description: Production
  - url: http://localhost:8787
    description: Development
paths:
...
  /tasks:
    get:
      tags:
        - Tasks
      summary: Listar tareas
      operationId: listTasks
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Lista de tareas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      tags:
        - Tasks
      summary: Crear nueva tarea
      operationId: createTask
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreate"
      responses:
        "201":
          description: Tarea creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
  /tasks/{taskId}:
    get:
      tags:
        - Tasks
      summary: Obtener detalles de tarea
      operationId: getTask
      security:
        - BearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Detalles de la tarea
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TaskCreate:
      type: object
      required:
        - title
        - mode
      properties:
        title:
          type: string
        description:
          type: string
        mode:
          type: string
          enum: [broadcast, targeted, auto]
        assigned_agents:
          type: array
          items:
            type: string
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, assigned, in_progress, completed, failed]
        mode:
          type: string
        assigned_agents:
          type: array
          items:
            type: string
        result:
          type: object
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
...